hooked various methods and created factory (and impl) for constructors

+++ vibrant-interface/src/main/kotlin/net/cydhra/vibrant/api/entity/VibrantPlayerSP.kt
@@ -20,5 +20,5 @@ interface VibrantPlayerSP : VibrantEntityLiving {
 
     fun getHeldItem(): VibrantItemStack?
 
-    fun getEyeHeight(): Double
+    fun getEyeHeight(): Float
 }
